+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is derived from [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- `changelog-md add [added|changed|deprecated|removed|fixed|security] [description]`
+
+
 ## 1.0.0 - 2025-02-24
 
 Initial release, including the 1.0 Schema specification